Remove API-specific RenderSurface class headers and RenderSurface code duplication.
